backend/synchronization/api/serializers.py
from rest_framework import serializers

from synchronization.models import Order


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = '__all__'
