backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
flake8==4.0.1
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
kombu==5.2.4
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.2.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==3.5.3
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
